Starting to build chart config factory
