+++ src/main/java/com/royware/corona/dashboard/interfaces/charts/IChartConfigFactory.java
@@ -1,0 +1,9 @@
+package com.royware.corona.dashboard.interfaces.charts;
+
+import org.springframework.stereotype.Service;
+import com.royware.corona.dashboard.enums.charts.ChartTypes;
+
+@Service
+public interface IChartConfigFactory {
+	public IChartConfigMaker create(ChartTypes chartType);
+}
