+++ src/main/java/com/royware/corona/dashboard/interfaces/charts/IChartConfigMaker.java
@@ -1,0 +1,13 @@
+package com.royware.corona.dashboard.interfaces.charts;
+
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.stereotype.Service;
+
+import com.royware.corona.dashboard.model.dashboard.DashboardChartConfig;
+
+@Service
+public interface IChartConfigMaker {
+	public DashboardChartConfig makeConfigFrom(List<List<Map<Object, Object>>> chartList, String forRegion);
+}
