+++ src/main/java/com/royware/corona/dashboard/services/chart/config/factory/ChartConfigStore.java
@@ -1,0 +1,27 @@
+package com.royware.corona.dashboard.services.chart.config.factory;
+
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import com.royware.corona.dashboard.enums.charts.ChartTypes;
+import com.royware.corona.dashboard.interfaces.charts.IChartListFactory;
+import com.royware.corona.dashboard.interfaces.charts.IChartListStore;
+import com.royware.corona.dashboard.interfaces.model.CanonicalCaseDeathData;
+
+@Component
+public class ChartConfigStore implements IChartListStore {
+	@Autowired
+	private IChartListFactory chartListFactory;
+	
+	@Override
+	public <T extends CanonicalCaseDeathData> List<List<Map<Object, Object>>> produceChartListFromRegionData(
+			ChartTypes chartType,
+			List<T> regionData,
+			int regionPopulation) {
+		
+		return chartListFactory.create(chartType).makeListFrom(regionData, regionPopulation);
+	}	
+}
