+++ src/main/java/com/royware/corona/dashboard/services/chart/list/makers/DailyAndTotalCasesVersusTimeChartList.java
@@ -1,4 +1,4 @@
-package com.royware.corona.dashboard.services.chartlistmakers;
+package com.royware.corona.dashboard.services.chart.list.makers;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -14,27 +14,24 @@ import com.royware.corona.dashboard.interfaces.charts.IChartListMaker;
 import com.royware.corona.dashboard.interfaces.model.CanonicalCaseDeathData;
 
 @Component
-public class TotalDeathsVersusTimeWithExponentialFitChartList implements IChartListMaker {
-	private static final Logger log = LoggerFactory.getLogger(TotalDeathsVersusTimeWithExponentialFitChartList.class);
-	private Map<Integer, Double> dailyDeaths = new HashMap<>();
+public class DailyAndTotalCasesVersusTimeChartList implements IChartListMaker {
+	private static final Logger log = LoggerFactory.getLogger(DailyAndTotalCasesVersusTimeChartList.class);
+	private Map<Integer, Double> dailyNewCases = new HashMap<>();
 
 	@Override
 	public <T extends CanonicalCaseDeathData> List<List<Map<Object, Object>>> makeListFrom(List<T> regionDataList, int pop) {
-		log.debug("MAKING TOTAL AND DAILY DEATHS VERSUS TIME");
+		log.debug("MAKING TOTAL AND DAILY CASES VERSUS TIME");
 		//Transform the data into ChartJS-ready lists
 		Map<Object, Object> xyPair;
 		List<Map<Object, Object>> dataList = new ArrayList<>();
 		List<List<Map<Object, Object>>> scatterChartDataLists = new ArrayList<>();
-
-		log.debug("Making time history of TOTAL deaths");
-		log.debug("Making time history of DAILY deaths");
-		int startDayIndex = ChartListMakerUtilities.findFirstDayIndexWithPositiveDeaths(regionDataList);
-		xyPair = ChartListMakerUtilities.makeXYPairWithDateStamp(startDayIndex,
-				regionDataList.get(startDayIndex).getTotalDeaths(),
-				regionDataList.get(startDayIndex).getDateChecked().toString());
+		
+		//First day
+		xyPair = ChartListMakerUtilities.makeXYPairWithDateStamp(0, regionDataList.get(0).getTotalPositiveCases(), regionDataList.get(0).getDateChecked().toString());
 		dataList.add(xyPair);
 		
-		int dayIndex = startDayIndex + 1;
+		//All subsequent days
+		int dayIndex = 1;
 		int totalYesterday = 0;
 		int totalToday = 0;
 		int dailyChange = 0;
@@ -42,31 +39,30 @@ public class TotalDeathsVersusTimeWithExponentialFitChartList implements IChartL
 			//TOTAL
 			xyPair = ChartListMakerUtilities.makeXYPairWithDateStamp(
 					dayIndex,
-					regionDataList.get(dayIndex).getTotalDeaths(),
+					regionDataList.get(dayIndex).getTotalPositiveCases(),
 					regionDataList.get(dayIndex).getDateChecked().toString());
 			dataList.add(xyPair);
 			
 			//NEW
-			totalYesterday = regionDataList.get(dayIndex - 1).getTotalDeaths();
-			totalToday = regionDataList.get(dayIndex).getTotalDeaths();
+			totalYesterday = regionDataList.get(dayIndex - 1).getTotalPositiveCases();
+			totalToday = regionDataList.get(dayIndex).getTotalPositiveCases();
 			dailyChange = totalToday - totalYesterday;
 			dailyChange = dailyChange > 0 ? dailyChange : 0;
-			dailyDeaths.put(dayIndex, dailyChange * 1.0);
-
+			dailyNewCases.put(dayIndex, dailyChange * 1.0);
+			
 			dayIndex++;
 		}
 		scatterChartDataLists.add(dataList);
 
-		//make the MOVING AVERAGE of daily deaths to smooth out some of the noise
-		log.debug("Making moving average of DAILY deaths");
+		//make the MOVING AVERAGE of daily quantity to smooth out some of the noise
 		scatterChartDataLists.add(
 			ChartListMakerUtilities.makeMovingAverageList(
-				dailyDeaths,
-				startDayIndex + MovingAverageSizes.MOVING_AVERAGE_SIZE.getValue(),
-				regionDataList.size()
+					dailyNewCases,
+					MovingAverageSizes.MOVING_AVERAGE_SIZE.getValue(),
+					regionDataList.size()
 			));
 		
-		log.debug("DONE MAKING TOTAL AND DAILY DEATHS VERSUS TIME");
+		log.debug("DONE MAKING TOTAL AND DAILY CASES VERSUS TIME");
 
 		return scatterChartDataLists;
 	}
