+++ src/main/java/com/royware/corona/dashboard/services/chart/list/makers/DailyRateOfChangeOfDeathsWithMovingAverageChartList.java
@@ -1,7 +1,8 @@
-package com.royware.corona.dashboard.services.chartlistmakers;
+package com.royware.corona.dashboard.services.chart.list.makers;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -14,35 +15,35 @@ import com.royware.corona.dashboard.interfaces.charts.IChartListMaker;
 import com.royware.corona.dashboard.interfaces.model.CanonicalCaseDeathData;
 
 @Component
-public class DailyAccelerationOfDeathsWithMovingAverageChartList implements IChartListMaker {
-	private static final Logger log = LoggerFactory.getLogger(DailyAccelerationOfDeathsWithMovingAverageChartList.class);
+public class DailyRateOfChangeOfDeathsWithMovingAverageChartList implements IChartListMaker {
+	private static final Logger log = LoggerFactory.getLogger(DailyRateOfChangeOfDeathsWithMovingAverageChartList.class);
+	private Map<Integer, Double> dailyPctChgDeaths = new LinkedHashMap<>();
 
 	@Override
 	public <T extends CanonicalCaseDeathData> List<List<Map<Object, Object>>> makeListFrom(List<T> regionDataList, int pop) {
-		log.debug("MAKING ACCELERATION OF DAILY DEATHS VERSUS TIME");
+		log.debug("MAKING RATE OF CHANGE OF DAILY DEATHS VERSUS TIME");
 		//Transform the data into ChartJS-ready lists
 		Map<Object, Object> xyPair;
 		List<Map<Object, Object>> dataList = new ArrayList<>();
 		List<List<Map<Object, Object>>> scatterChartDataLists = new ArrayList<>();
-		Map<Integer, Double> dailyAccelDeaths = new HashMap<>();
 		
-		double valueToday;
-		double valueYesterday;
-		double valueDayBeforeYesterday;
-		double accelerationOfDeaths = 0;
-		int startDayIndex = ChartListMakerUtilities.findFirstDayIndexWithPositiveDeaths(regionDataList);
-		int dayIndex = startDayIndex + 2;  //day 2 in list
+		dailyPctChgDeaths.clear();
+		
+		int startDayIndex = ChartListMakerUtilities.findFirstDayIndexWithPositiveDeaths(regionDataList) + 1;
+		log.debug("startDayIndex for daily deaths: " + startDayIndex);
+		int valueToday;
+		int valueYesterday;
+		double percentChange = 0;
+		int dayIndex = startDayIndex;
 		while(dayIndex < regionDataList.size()) {
-			valueDayBeforeYesterday = regionDataList.get(dayIndex - 2).getTotalDeaths();
 			valueYesterday = regionDataList.get(dayIndex - 1).getTotalDeaths();
 			valueToday = regionDataList.get(dayIndex).getTotalDeaths();
-			accelerationOfDeaths = ((valueToday - valueYesterday) - (valueYesterday - valueDayBeforeYesterday)) * 100.0 / (valueToday - valueYesterday);
-			
-			dailyAccelDeaths.put(dayIndex, accelerationOfDeaths);
+			percentChange = (valueToday - valueYesterday) * 100.0 / valueYesterday;
+			dailyPctChgDeaths.put(dayIndex, percentChange);
 			
 			xyPair = new HashMap<>();
 			xyPair.put("x", dayIndex);
-			xyPair.put("y", accelerationOfDeaths);
+			xyPair.put("y", percentChange);
 			xyPair.put("dateChecked", regionDataList.get(dayIndex).getDateChecked().toString());
 			dataList.add(xyPair);
 			
@@ -53,12 +54,12 @@ public class DailyAccelerationOfDeathsWithMovingAverageChartList implements ICha
 		//make the MOVING AVERAGE
 		scatterChartDataLists.add(
 				ChartListMakerUtilities.makeMovingAverageList(
-					dailyAccelDeaths,
-					startDayIndex + MovingAverageSizes.MOVING_AVERAGE_SIZE.getValue() + 1,
+					dailyPctChgDeaths,
+					startDayIndex + MovingAverageSizes.MOVING_AVERAGE_SIZE.getValue(),
 					regionDataList.size()
 				));
 
-		log.debug("DONE MAKING ACCELERATION OF DAILY DEATHS VERSUS TIME");
+		log.debug("DONE MAKING RATE OF CHANGE OF DAILY DEATHS VERSUS TIME");
 
 		return scatterChartDataLists;
 	}
